refactor(sidebar): extract GroupRow, SessionRow, NewSessionButton (Task #723 Phase A)
SRP split: 1033 LoC bag -> 4 single-feature modules.

- src/components/sidebar/GroupRow.tsx (291 LoC) — group header + sortable list
- src/components/sidebar/SessionRow.tsx (295 LoC) — session row + context menu
- src/components/sidebar/NewSessionButton.tsx (64 LoC) — new-session pill cluster
- src/components/sidebar/dnd.ts (12 LoC) — header droppable id helpers
- src/components/Sidebar.tsx slimmed 1033 -> 426 LoC; keeps DndContext + layout
  + scroll/keyboard handling + collapsed rail.

Tests:
- New tests/sidebar/GroupRow.test.tsx (2 cases) — label/count/aria-expanded
  when expanded; list hidden + aria-expanded=false when collapsed.
- New tests/sidebar/SessionRow.test.tsx (2 cases) — aria-listbox option +
  data-session-id; crashed indicator surfaces from disconnectedSessions store.
- Existing tests/sidebar-resizer-keyboard.test.tsx (13 cases) still PASS.

Verified:
- npx tsc --noEmit -> clean (sidebar files only; pre-existing app-effects
  TS error unrelated and not introduced by this PR).
- npx tsc --noEmit -p tsconfig.electron.json -> clean.
- npx vitest run tests/sidebar tests/sidebar-resizer-keyboard.test.tsx
  -> 17/17 PASS (3 files).
- Full vitest (excluding pre-existing db-hardening native ABI flake +
  pre-existing app-effects unrelated changes) -> 515/515 PASS.
- npx webpack --mode production -> build OK.
