+++ src/components/Sidebar.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { motion } from 'framer-motion';
 import {
-  ChevronDown,
   ChevronRight,
   Download,
   Plus,
@@ -13,18 +12,11 @@ import {
   DragOverlay,
   PointerSensor,
   closestCenter,
-  useDroppable,
   useSensor,
   useSensors,
   type DragEndEvent,
   type DragStartEvent
 } from '@dnd-kit/core';
-import {
-  SortableContext,
-  useSortable,
-  verticalListSortingStrategy
-} from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
 import { useStore } from '../stores/store';
 import { cn } from '../lib/cn';
 import { IconButton } from './ui/IconButton';
@@ -32,560 +24,19 @@ import { Button } from './ui/Button';
 import { AgentIcon } from './AgentIcon';
 import { CwdPopover } from './CwdPopover';
 import { DragRegion } from './WindowControls';
-import {
-  ContextMenu,
-  ContextMenuContent,
-  ContextMenuItem,
-  ContextMenuSeparator,
-  ContextMenuSub,
-  ContextMenuSubContent,
-  ContextMenuSubTrigger,
-  ContextMenuTrigger
-} from './ui/ContextMenu';
-import { InlineRename } from './ui/InlineRename';
-import { ConfirmDialog } from './ui/ConfirmDialog';
-import { useToast } from './ui/Toast';
 import { useTranslation } from '../i18n/useTranslation';
 import {
   DURATION_RAW,
   EASING,
-  MOTION_SESSION_SWITCH_DURATION,
-  MOTION_STANDARD_EASING,
 } from '../lib/motion';
-import type { Group, Session } from '../types';
+import type { Session } from '../types';
+import { GroupRow } from './sidebar/GroupRow';
+import { NewSessionButton } from './sidebar/NewSessionButton';
+import { parseHeaderDroppable } from './sidebar/dnd';
 
 // Session order inside a group is user-controlled (drag to reorder) — not
 // derived from state. The array order handed down from the store is the
 // source of truth; don't re-sort it here.
-
-// Cross-group DnD uses three drop-target id flavors via closestCenter:
-//   - session id            → insert before that session
-//   - group id              → append to empty SortableContext (only when open)
-//   - `header:<groupId>`    → append to that group (works even if collapsed,
-//                             and drives hover-to-expand on collapsed groups)
-const HEADER_PREFIX = 'header:';
-const headerDroppableId = (groupId: string) => `${HEADER_PREFIX}${groupId}`;
-const parseHeaderDroppable = (id: string) =>
-  id.startsWith(HEADER_PREFIX) ? id.slice(HEADER_PREFIX.length) : null;
-
-function GroupRow({
-  group,
-  sessions,
-  activeSessionId,
-  focused,
-  anyGroupFocused,
-  autoRename,
-  onSelectSession,
-  onFocus,
-  normalGroups
-}: {
-  group: Group;
-  sessions: Session[];
-  activeSessionId: string;
-  focused: boolean;
-  anyGroupFocused: boolean;
-  /** When true, the GroupRow mounts in inline-rename mode with the input
-   *  pre-focused — used right after `createGroup()` so the user can name the
-   *  group without an extra click. Cleared by the parent once consumed. */
-  autoRename?: boolean;
-  onSelectSession: (id: string) => void;
-  onFocus: () => void;
-  /** Pre-filtered list of normal (non-archive) groups, hoisted to the parent
-   *  Sidebar so we don't recompute per SessionRow per render. */
-  normalGroups: Group[];
-}) {
-  const { t } = useTranslation();
-  const sessionIds = sessions.map((s) => s.id);
-  const hasWaiting = sessions.some((s) => s.state === 'waiting');
-  const setGroupCollapsed = useStore((s) => s.setGroupCollapsed);
-  const renameGroup = useStore((s) => s.renameGroup);
-  const deleteGroup = useStore((s) => s.deleteGroup);
-  const restoreGroup = useStore((s) => s.restoreGroup);
-  const archiveGroup = useStore((s) => s.archiveGroup);
-  const unarchiveGroup = useStore((s) => s.unarchiveGroup);
-  const toast = useToast();
-  const collapsed = group.collapsed;
-  const [renaming, setRenaming] = useState(!!autoRename);
-  const [confirmDelete, setConfirmDelete] = useState(false);
-  const isSpecial = group.kind !== 'normal';
-  // Previously gated against a `kind: 'deleted'` enum value that no code path
-  // ever produced. Now that the enum is normal | archive only, no group is
-  // ever menu-disabled (archived groups still allow rename / unarchive).
-  const menuDisabled = false;
-  const { isOver, setNodeRef: setHeaderDropRef } = useDroppable({
-    id: headerDroppableId(group.id),
-    data: { type: 'group-header', groupId: group.id }
-  });
-  // Hover-to-expand: if a session is dragged over a collapsed group's header
-  // for 400ms, open it so the user can drop further into its list. Cancel if
-  // they leave before the timer fires.
-  const expandTimerRef = useRef<number | null>(null);
-  useEffect(() => {
-    if (!isOver || !collapsed) {
-      if (expandTimerRef.current !== null) {
-        window.clearTimeout(expandTimerRef.current);
-        expandTimerRef.current = null;
-      }
-      return;
-    }
-    expandTimerRef.current = window.setTimeout(() => {
-      setGroupCollapsed(group.id, false);
-      expandTimerRef.current = null;
-    }, 400);
-    return () => {
-      if (expandTimerRef.current !== null) {
-        window.clearTimeout(expandTimerRef.current);
-        expandTimerRef.current = null;
-      }
-    };
-  }, [isOver, collapsed, group.id, setGroupCollapsed]);
-  return (
-    <div className="mb-2">
-      <ContextMenu>
-        <ContextMenuTrigger asChild disabled={menuDisabled}>
-          <div
-            ref={setHeaderDropRef}
-            data-group-header-id={group.id}
-            className={cn(
-              'group/row relative flex items-center h-7 px-2 rounded-sm transition-colors duration-120 ease-out',
-              focused
-                ? 'bg-bg-active'
-                : 'hover:bg-bg-hover',
-              isOver && 'ring-1 ring-inset ring-accent bg-bg-active'
-            )}
-          >
-            {focused && (
-              <motion.span
-                aria-hidden
-                initial={{ scaleY: 0, opacity: 0, x: -2 }}
-                animate={{ scaleY: 1, opacity: 1, x: 0 }}
-                transition={{ duration: DURATION_RAW.ms220, ease: EASING.standard }}
-                style={{ originY: 0.5 }}
-                className="absolute left-0 top-0 bottom-0 w-[3px] bg-accent rounded-r-sm"
-              />
-            )}
-            <button
-              onClick={() => {
-                onFocus();
-                if (!renaming) setGroupCollapsed(group.id, !collapsed);
-              }}
-              onKeyDown={(e) => {
-                // F2 enters rename mode — matches the SessionRow F2 handler
-                // and Finder / VS Code explorer convention. Special groups
-                // (archive) have no rename action in their menu, so don't
-                // permit F2 either to keep behavior consistent with the
-                // visible affordances. (`isSpecial` is checked via the same
-                // `kind !== 'normal'` rule used for the menu disable gate.)
-                if (e.key === 'F2' && !renaming && !isSpecial) {
-                  e.preventDefault();
-                  setRenaming(true);
-                }
-              }}
-              className="flex flex-1 min-w-0 items-center gap-1.5 text-left text-fg-secondary outline-none rounded-sm focus-ring"
-              aria-expanded={!collapsed}
-            >
-              <motion.span
-                initial={false}
-                animate={{ rotate: collapsed ? 0 : 90 }}
-                transition={{ duration: DURATION_RAW.ms200, ease: EASING.enter }}
-                className="inline-flex shrink-0"
-              >
-                <ChevronRight size={12} className="stroke-[1.75] text-fg-tertiary" />
-              </motion.span>
-              {renaming ? (
-                <InlineRename
-                  value={group.name}
-                  onCommit={(next) => {
-                    renameGroup(group.id, next);
-                    setRenaming(false);
-                  }}
-                  onCancel={() => setRenaming(false)}
-                  inputClassName="text-chrome font-semibold text-fg-primary"
-                />
-              ) : (
-                <>
-                  <span
-                    className="truncate text-chrome font-semibold text-fg-secondary"
-                    onDoubleClick={(e) => {
-                      // Double-click the label to enter rename mode — matches
-                      // the SessionRow handler. Stop propagation so the
-                      // wrapping <button>'s onClick doesn't toggle collapsed
-                      // on the same gesture. Skip for special (archive)
-                      // groups since they have no rename action.
-                      if (isSpecial) return;
-                      e.stopPropagation();
-                      setRenaming(true);
-                    }}
-                  >{group.nameKey ? t(group.nameKey) : group.name}</span>
-                  {hasWaiting && (
-                    <span
-                      aria-label={t('sidebar.waitingForResponse')}
-                      className="ml-1.5 shrink-0 inline-block w-1.5 h-1.5 rounded-full bg-state-waiting"
-                    />
-                  )}
-                  {!isSpecial && sessions.length > 0 && (
-                    <span className="ml-1 text-label-meta">{sessions.length}</span>
-                  )}
-                </>
-              )}
-            </button>
-          </div>
-        </ContextMenuTrigger>
-        <ContextMenuContent
-          // Prevent Radix from restoring focus to the trigger after the
-          // menu closes. When `Rename` is selected, Radix would otherwise
-          // .focus() the trigger element AFTER our InlineRename mount
-          // effect, racing the input's focus. Group rows happen to win
-          // that race today (no dnd-kit listeners on the trigger), but
-          // belt-and-suspenders here keeps parity with SessionRow and
-          // hardens against future regressions.
-          onCloseAutoFocus={(e) => e.preventDefault()}
-        >
-          <ContextMenuItem onSelect={() => setRenaming(true)}>{t('common.rename')}</ContextMenuItem>
-          <ContextMenuSeparator />
-          <ContextMenuItem
-            onSelect={() =>
-              group.kind === 'archive' ? unarchiveGroup(group.id) : archiveGroup(group.id)
-            }
-          >
-            {group.kind === 'archive' ? t('sidebar.unarchiveGroup') : t('sidebar.archiveGroup')}
-          </ContextMenuItem>
-          <ContextMenuItem danger onSelect={() => setConfirmDelete(true)}>
-            {t('sidebar.deleteGroupEllipsis')}
-          </ContextMenuItem>
-        </ContextMenuContent>
-      </ContextMenu>
-      {!collapsed && (
-        <SortableContext items={sessionIds} strategy={verticalListSortingStrategy} id={group.id}>
-          <ul
-            className="mt-px"
-            data-group-id={group.id}
-            role="listbox"
-          aria-label={group.name}
-          onKeyDown={(e) => {
-            // Don't hijack arrows while the inline-rename input is focused —
-            // the user is editing text and expects ←/→ to move the caret,
-            // ↑/↓ to be no-ops (browsers don't move caret vertically in a
-            // single-line input but also don't navigate the listbox). Same
-            // guard pattern lives on the SessionRow's own onKeyDown.
-            if ((e.target as HTMLElement).tagName === 'INPUT') return;
-            const key = e.key;
-            if (key !== 'ArrowDown' && key !== 'ArrowUp' && key !== 'Home' && key !== 'End') return;
-            const list = e.currentTarget;
-            const items = Array.from(
-              list.querySelectorAll<HTMLLIElement>('li[role="option"]')
-            );
-            if (items.length === 0) return;
-            const current = document.activeElement as HTMLElement | null;
-            const idx = current ? items.indexOf(current as HTMLLIElement) : -1;
-            let next = idx;
-            if (key === 'ArrowDown') next = idx < 0 ? 0 : Math.min(idx + 1, items.length - 1);
-            else if (key === 'ArrowUp') next = idx < 0 ? items.length - 1 : Math.max(idx - 1, 0);
-            else if (key === 'Home') next = 0;
-            else if (key === 'End') next = items.length - 1;
-            if (next !== idx && items[next]) {
-              e.preventDefault();
-              items[next].focus();
-            }
-          }}
-        >
-          {sessions.map((s) => (
-            <SessionRow
-              key={s.id}
-              session={s}
-              active={s.id === activeSessionId}
-              selected={!anyGroupFocused && s.id === activeSessionId}
-              onSelect={() => onSelectSession(s.id)}
-              normalGroups={normalGroups}
-            />
-          ))}
-          </ul>
-        </SortableContext>
-      )}
-      <ConfirmDialog
-        open={confirmDelete}
-        onOpenChange={setConfirmDelete}
-        title={t('sidebar.deleteGroupConfirmTitle', { name: group.name })}
-        description={
-          sessions.length > 0
-            ? t('sidebar.deleteGroupNonEmptyKept', { count: sessions.length })
-            : t('sidebar.deleteGroupEmpty')
-        }
-        confirmLabel={t('sidebar.deleteGroup')}
-        cancelLabel={t('common.cancel')}
-        destructive
-        onConfirm={() => {
-          const snap = deleteGroup(group.id);
-          if (snap) {
-            toast.push({
-              kind: 'info',
-              title: t('sidebar.groupDeletedToast', { name: snap.group.name }),
-              action: { label: t('common.undo'), onClick: () => restoreGroup(snap) }
-            });
-          }
-        }}
-      />
-    </div>
-  );
-}
-
-function SessionRow({ session, active, selected, onSelect, normalGroups }: { session: Session; active: boolean; selected: boolean; onSelect: () => void; normalGroups: Group[] }) {
-  const { t } = useTranslation();
-  const [renaming, setRenaming] = useState(false);
-  // Transient flash from the notify pipeline (#689). ORed into AgentIcon's
-  // halo trigger via the `flashing` prop so Rule 2 short-task pulses appear
-  // even when `state === 'idle'`.
-  const flashing = useStore((s) => s.flashStates[session.id] === true);
-  // Per-session pty disconnect classification — populated by the app-boot
-  // unconditional `pty:exit` listener (App.tsx). We surface the red dot
-  // ONLY for `crashed` (signal or non-zero exit) — `clean` exits are
-  // user-intentional (typed `/exit`) and shouldn't grab attention. The
-  // dot clears when the user retries (TerminalPane → `_clearPtyExit`).
-  const crashed = useStore(
-    (s) => s.disconnectedSessions[session.id]?.kind === 'crashed'
-  );
-  // Perf: receive `normalGroups` as a prop computed once in the parent
-  // <Sidebar>, rather than calling `s.groups.filter(...)` per row per render.
-  const groups = normalGroups;
-  const renameSession = useStore((s) => s.renameSession);
-  const deleteSession = useStore((s) => s.deleteSession);
-  const restoreSession = useStore((s) => s.restoreSession);
-  const moveSession = useStore((s) => s.moveSession);
-  const createGroup = useStore((s) => s.createGroup);
-  const toast = useToast();
-  const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
-    id: session.id,
-    data: { type: 'session', groupId: session.groupId }
-  });
-  const liRef = useRef<HTMLLIElement | null>(null);
-  // Compose dnd-kit's ref with our own so we can call scrollIntoView when the
-  // active session changes from off-screen → selected (J17).
-  const composedRef = (node: HTMLLIElement | null) => {
-    liRef.current = node;
-    setNodeRef(node);
-  };
-  // J17: when this row becomes the active selection (e.g. via palette,
-  // notification click, programmatic select), bring it into the sidebar
-  // viewport. `block: 'nearest'` avoids snapping when the row is already
-  // visible. Skip during drag so we don't fight the dnd transform.
-  useEffect(() => {
-    if (!selected || isDragging) return;
-    const el = liRef.current;
-    if (!el) return;
-    el.scrollIntoView({ block: 'nearest', behavior: 'smooth' });
-  }, [selected, isDragging]);
-  const style: React.CSSProperties = {
-    transform: isDragging ? undefined : CSS.Transform.toString(transform),
-    transition,
-    opacity: isDragging ? 0 : 1
-  };
-  // J4: session delete is a soft-delete with undo toast. We removed the
-  // ConfirmDialog gate — destructive intent is already obvious from the
-  // menu label, and the toast lets the user reverse a misclick within ~5s.
-  const performDelete = () => {
-    const snap = deleteSession(session.id);
-    if (snap) {
-      toast.push({
-        kind: 'info',
-        title: t('sidebar.sessionDeletedToast', { name: snap.session.name }),
-        action: { label: t('common.undo'), onClick: () => restoreSession(snap) }
-      });
-    }
-  };
-  return (
-    <ContextMenu>
-      <ContextMenuTrigger asChild>
-        <li
-          ref={composedRef}
-          style={style}
-          {...attributes}
-          // Skip dnd-kit pointer listeners and remove tabIndex while the
-          // inline-rename input is mounted. Two reasons:
-          //   1. Drag-during-rename has no useful semantics.
-          //   2. The listeners `preventDefault()` on mousedown, which
-          //      combined with `tabIndex=0` lets the LI win the focus
-          //      race against the just-mounted input — defeats Fix A1's
-          //      onCloseAutoFocus guard if Radix dispatches a synthetic
-          //      mouse event during close. Belt-and-suspenders for the
-          //      session focus race the user reported.
-          {...(renaming ? {} : listeners)}
-          role="option"
-          aria-selected={selected}
-          tabIndex={renaming ? -1 : selected ? 0 : -1}
-          data-session-id={session.id}
-          onClick={onSelect}
-          onContextMenu={() => {
-            // J4b: right-click selects the row first, matching standard GUI
-            // behavior where the context menu acts on "this row" — and the
-            // user expects "this row" to be visually highlighted.
-            onSelect();
-          }}
-          onKeyDown={(e) => {
-            // Only handle keys when the <li> itself is the focused element.
-            // Without this guard, typing a space in the inline-rename <input>
-            // bubbles up here and gets preventDefault()-ed → spaces silently
-            // disappear from the new name. Same risk for Enter, which we want
-            // the input to commit on.
-            if (e.target !== e.currentTarget) return;
-            // F2 enters rename mode on the focused row — matches Finder /
-            // VS Code explorer / Windows Explorer convention.
-            if (e.key === 'F2' && !renaming) {
-              e.preventDefault();
-              setRenaming(true);
-              return;
-            }
-            if (e.key === 'Enter' || e.key === ' ') {
-              e.preventDefault();
-              onSelect();
-            }
-          }}
-          title={
-            session.cwdMissing
-              ? t('sidebar.cwdMissingTooltip', { cwd: session.cwd })
-              : undefined
-          }
-          className={cn(
-            'group/sess relative flex items-center gap-2.5 pl-3 pr-2 rounded-sm cursor-pointer text-chrome h-9',
-            // Keep duration/easing in sync with the selection ring below and
-            // with the right-pane content crossfade in ChatStream so clicking
-            // a session reads as ONE coordinated motion across panes. See
-            // src/lib/motion.ts (MOTION_SESSION_SWITCH_DURATION / EASING).
-            'transition-[background-color,color,box-shadow] duration-[180ms]',
-            '[transition-timing-function:cubic-bezier(0.32,0.72,0,1)]',
-            'outline-none focus-ring',
-            selected
-              ? 'bg-bg-active text-fg-primary'
-              : 'text-fg-secondary hover:bg-bg-hover hover:text-fg-primary',
-            active && 'font-medium text-fg-primary',
-            session.cwdMissing && 'opacity-55'
-          )}
-        >
-          {selected && (
-            <motion.span
-              aria-hidden
-              initial={{ scaleY: 0, opacity: 0, x: -2 }}
-              animate={{ scaleY: 1, opacity: 1, x: 0 }}
-              transition={{
-                duration: MOTION_SESSION_SWITCH_DURATION,
-                ease: MOTION_STANDARD_EASING
-              }}
-              style={{ originY: 0.5 }}
-              className="absolute left-0 top-0 bottom-0 w-[3px] bg-accent rounded-r-sm"
-            />
-          )}
-          <AgentIcon agentType={session.agentType} state={session.state} flashing={flashing} size="sm" />
-          <span className="flex-1 min-w-0 leading-tight block">
-            {renaming ? (
-              <InlineRename
-                value={session.name}
-                onCommit={(next) => {
-                  // renameSession is now async (optimistic update + SDK
-                  // writeback); we don't await — local name updates
-                  // synchronously and the JSONL writeback runs in the
-                  // background. Errors are logged, not toasted.
-                  void renameSession(session.id, next);
-                  setRenaming(false);
-                }}
-                onCancel={() => setRenaming(false)}
-                inputClassName="text-chrome"
-              />
-            ) : (
-              <>
-                <span
-                  className="truncate block"
-                  title={session.name}
-                  onDoubleClick={(e) => {
-                    // Double-click the label to enter rename mode — matches
-                    // Finder / VS Code explorer convention. Stop propagation
-                    // so the row's onClick doesn't fire a redundant select.
-                    e.stopPropagation();
-                    setRenaming(true);
-                  }}
-                >{session.name}</span>
-              </>
-            )}
-          </span>
-          {active && (
-            <span className="sidebar-rail-cell sidebar-rail-cell--nested shrink-0">
-              <span
-                aria-label={t('sidebar.openInChat')}
-                className="shrink-0 inline-block w-1.5 h-1.5 rounded-full bg-accent"
-              />
-            </span>
-          )}
-          {crashed && (
-            // Crash indicator — sits to the right of the existing
-            // state pip (separate rail cell so it doesn't displace
-            // the live-state dot when both apply, e.g. a previously-
-            // running session that just died). Reuses the canonical
-            // `bg-state-error` token (no new red shade introduced).
-            <span
-              className="sidebar-rail-cell sidebar-rail-cell--nested shrink-0"
-              data-session-crashed
-              title={t('sidebar.sessionCrashed')}
-            >
-              <span
-                aria-label={t('sidebar.sessionCrashed')}
-                className="shrink-0 inline-block w-1.5 h-1.5 rounded-full bg-state-error"
-              />
-            </span>
-          )}
-        </li>
-      </ContextMenuTrigger>
-      <ContextMenuContent
-        // Prevent Radix from restoring focus to the LI trigger after the
-        // menu closes. Without this, Radix's `onCloseAutoFocus` calls
-        // .focus() on the LI AFTER our InlineRename mount effect, and
-        // because the LI carries dnd-kit `{...listeners}` plus
-        // `tabIndex={selected ? 0 : -1}` it wins the focus race against
-        // the just-mounted input. The user then types and nothing happens
-        // because keys land on the (focused) LI, not the input. PR #527
-        // fixed the auto-cancel side of this race; this preventDefault
-        // closes the focus-stealing side.
-        onCloseAutoFocus={(e) => e.preventDefault()}
-      >
-        <ContextMenuItem onSelect={() => setRenaming(true)}>{t('common.rename')}</ContextMenuItem>
-        {(() => {
-          // Exclude the session's current group from the destination list so
-          // users can only move TO another group (#612). The submenu is
-          // ALWAYS rendered so users can still create a new group via the
-          // "New group..." escape hatch even when no other group exists (#629).
-          const otherGroups = groups.filter((g) => g.id !== session.groupId);
-          return (
-            <ContextMenuSub>
-              <ContextMenuSubTrigger data-testid="move-to-group-trigger">{t('sidebar.moveToGroup')}</ContextMenuSubTrigger>
-              <ContextMenuSubContent data-testid="move-to-group-content">
-                {otherGroups.map((g) => (
-                  <ContextMenuItem
-                    key={g.id}
-                    data-move-to-group-item
-                    data-group-id={g.id}
-                    onSelect={() => moveSession(session.id, g.id, null)}
-                  >
-                    {g.name}
-                  </ContextMenuItem>
-                ))}
-                {otherGroups.length > 0 && <ContextMenuSeparator />}
-                <ContextMenuItem
-                  onSelect={() => {
-                    const id = createGroup();
-                    moveSession(session.id, id, null);
-                  }}
-                >
-                  {t('sidebar.newGroupEllipsis')}
-                </ContextMenuItem>
-              </ContextMenuSubContent>
-            </ContextMenuSub>
-          );
-        })()}
-        <ContextMenuSeparator />
-        <ContextMenuItem danger onSelect={performDelete}>
-          {t('common.delete')}
-        </ContextMenuItem>
-      </ContextMenuContent>
-    </ContextMenu>
-  );
-}
 
 export type SidebarProps = {
   /** Create a new session in-place. The store seeds `cwd` from the user's
@@ -603,64 +54,6 @@ export type SidebarProps = {
   sessions: Session[];
   onMoveSession: (sessionId: string, targetGroupId: string, beforeSessionId: string | null) => void;
 };
-
-function NewSessionButton({
-  onCreateSession,
-  cwdPopoverOpen,
-  onCwdPopoverOpenChange,
-  chevronRef
-}: {
-  onCreateSession?: () => void;
-  cwdPopoverOpen: boolean;
-  onCwdPopoverOpenChange: (open: boolean) => void;
-  chevronRef: React.RefObject<HTMLButtonElement>;
-}) {
-  const { t } = useTranslation();
-  // The "main" New Session button + "▾" cwd-picker chevron form a single
-  // visual cluster: shared raised pill with a hairline divider between the
-  // two halves. Each half has its own click target (44x44 hit padding via
-  // the buttons' own padding/grid), its own focus ring, and its own aria
-  // label. Tooltips help discoverability for the chevron specifically.
-  return (
-    <div className="flex-1 inline-flex items-stretch h-8" data-sidebar-newsession-cluster>
-      <Button
-        variant="raised"
-        size="md"
-        onClick={() => onCreateSession?.()}
-        className={cn(
-          'flex-1 h-8 text-chrome gap-1.5',
-          // Strip the right-side rounded corners so the chevron sits flush.
-          '!rounded-r-none border-r-0'
-        )}
-      >
-        <Plus size={14} className="stroke-[1.75]" />
-        <span>{t('sidebar.newSession')}</span>
-      </Button>
-      <IconButton
-        ref={chevronRef}
-        variant="raised"
-        size="md"
-        aria-label={t('sidebar.pickCwdAria')}
-        aria-expanded={cwdPopoverOpen}
-        aria-haspopup="dialog"
-        tooltip={t('sidebar.pickCwdTooltip')}
-        tooltipSide="bottom"
-        onClick={() => onCwdPopoverOpenChange(!cwdPopoverOpen)}
-        data-testid="sidebar-newsession-cwd-chevron"
-        className={cn(
-          'h-8 w-7 shrink-0',
-          // Match the cluster: kill left rounding so we sit flush with the
-          // main button's right edge, keep right rounding from the cluster
-          // wrapper. Subtle inner-left hairline divides the two halves.
-          '!rounded-l-none',
-          'shadow-[inset_1px_0_0_0_oklch(0_0_0_/_0.35)]'
-        )}
-      >
-        <ChevronDown size={12} className="stroke-[1.75]" />
-      </IconButton>
-    </div>
-  );
-}
 
 export function Sidebar({ onCreateSession, onOpenSettings, onOpenPalette, onOpenImport, activeSessionId, focusedGroupId, onSelectSession, onFocusGroup, sessions, onMoveSession }: SidebarProps) {
   const { t } = useTranslation();
